Finalize POC for integration-tests restructuring
