+++ integration-tests/modules/init.go
@@ -1,0 +1,7 @@
+package modules
+
+import integrationtests "github.com/CoreumFoundation/coreum/integration-tests"
+
+func init() {
+	integrationtests.Init()
+}
